requirements.txt
torch
numpy
streamlit
pinecone-client
torchvision
tqdm
pillow